Array indexes are unsigned now

+++ src/sodiumoxide/crypto/siphash24.rs
@@ -137,8 +137,8 @@ fn test_vectors() {
                       ,[0x57, 0x5f, 0xf2, 0x8e, 0x60, 0x38, 0x1b, 0xe5]
                       ,[0x72, 0x45, 0x06, 0xeb, 0x4c, 0x32, 0x8a, 0x95]];
     let k = Key([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15]);
-    for i in range(0, maxlen) {
-        let Digest(h) = shorthash(m.slice(0, i as uint), &k);
+    for i in range(0u, maxlen) {
+        let Digest(h) = shorthash(m.slice(0, i), &k);
         assert!(h == h_expecteds[i]);
     }
 }
